Add support for Pydantic models with `alias=`
This change ensures that the generated JSON schema
/ OpenAPI docs match the `alias=` param when
present, and that any returned JSON uses the alias
for the keys. This is the most consistent behavior
with Pydantic's usage, *but* is a breaking change
for any existing code that uses `alias=` and
relies on FastOpenAPI returning the field name
instead of the alias.

We could put this behind a setting if desired
instead, i.e.,:

```
model_dump(
    by_alias=(True if use_aliases else False)
)
model_json_schema(
    mode=("serialization" if use_aliases else "validation")
)
```